+++ Data/WrathClassic/ExportedTooltipData.lua
@@ -1,4 +1,4 @@
--- [[ Exported at 2023-04-13 13-01-25 ]] --
+-- [[ Exported at 2023-04-18 23-05-00 ]] --
 -- [[ This code is automatically generated as an export from ]] --
 -- [[ an SQLite database and is not meant for manual edit. ]] --