+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to this project will be documented in this file.
 ### Fixed
 - When GW2 UI skin is enable, the right click menu button (if shown) is now also skinned
 - When (un)tracking achievements while the achievement is visible in the achievement window, the tracked state of the achievement now also updated
+- Wildfire for Global Swarming now has the correct npc id
 
 ## 55.0 - 2023-04-01
 ### Added